app/requirements/Requirements.txt
# Web Framework
Flask==2.3.3
flask-talisman==1.1.0
datetime==5.1.0
requests==2.28.1
gunicorn==21.2.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
Flask-Migrate==4.0.5

# Security & Auth
flask-jwt-extended==4.5.3
Flask-Limiter==3.5.0
python-dotenv>=1.0.0
cryptography>=41.0.0

# Validation & API Docs
pydantic==2.4.2
flask-openapi3==3.0.1

# Caching
Flask-Caching==2.1.0
redis==5.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Monitoring
prometheus-flask-exporter==0.22.4

# Data Processing
# pandas==2.0.3

